requirements.txt
Flask~=3.1.0
Werkzeug~=3.1.3
Flask-JWT-Extended~=4.7.1
python-dotenv~=1.0.1
requests~=2.32.3
flask-sqlalchemy~=3.1.1
psycopg2-binary~=2.9.10
pylint~=3.3.4